fix: 로그아웃 로직 분리
